添加快捷键，修复BUG

+++ gui/GraphicBezier.cpp
@@ -18,7 +18,7 @@ void GraphicBezier::Serialize(CArchive & ar)
 {
 	if (ar.IsLoading())
 	{
-		int len;
+		long long int len;
 		ar >> len;
 		for (auto i = 0; i < len; i++)
 		{
