+++ gui/GraphicStyle.h
@@ -102,7 +102,7 @@ inline void GraphicAttrTimeline<T>::Serialize(CArchive & ar)
 {
 	if (ar.IsLoading())
 	{
-		int len;
+		long long int len;
 		ar >> len;
 		for (auto i = 0; i < len; i++)
 		{
